+++ examples/pissa_finetuning/pissa_finetuning.py
@@ -47,7 +47,7 @@ class TrainingArguments(TrainingArguments):
     # dataset configs
     data_path: str = field(default="imdb", metadata={"help": "Path to the training data."})
     dataset_split: str = field(
-        default="train[:1%]", metadata={"help": "init_lora_weights (`['train', 'test', 'eval']`):"}
+        default="train[:1%]", metadata={"help": "(`['train', 'test', 'eval']`):"}
     )
     dataset_field: List[str] = field(default=None, metadata={"help": "Fields of dataset input and output."})
     max_seq_length: int = field(
