+++ src/peft/tuners/lora/model.py
@@ -16,6 +16,7 @@ from __future__ import annotations
 import math
 import operator
 import re
+import os
 import warnings
 from contextlib import contextmanager
 from dataclasses import asdict, replace
@@ -42,6 +43,7 @@ from peft.utils import (
     _freeze_adapter,
     _get_submodules,
     get_quantization_config,
+    get_peft_model_state_dict
 )
 from peft.utils.merge_utils import dare_linear, dare_ties, magnitude_prune, task_arithmetic, ties
 
@@ -791,3 +793,24 @@ class LoraModel(BaseTuner):
         model.
         """
         return self._unload_and_optionally_merge(merge=False)
+
+    def subtract_pissa_init(self, pissa_initial_dir, output_state_dict, kwargs):
+        adapter_name = os.path.basename(pissa_initial_dir)
+        self.load_adapter(os.path.dirname(pissa_initial_dir), adapter_name=adapter_name, subfolder=adapter_name)
+        pissa_init_state_dict = get_peft_model_state_dict(
+            self,
+            state_dict=kwargs.get("state_dict", None),
+            adapter_name=adapter_name,
+        )
+        tensors_lora = {}
+        for name in pissa_init_state_dict.keys():
+            ## W = W^{res} + A_0 \times B_0,
+            ## W + \Delta W = W^{res} + A \times B,
+            ## \Delta W = A \times B - A_0 \times B_0 = [A | A_0] \times [B | B_0]^T = A'B'.
+            tensors_lora[name] = (
+                torch.cat([output_state_dict[name], pissa_init_state_dict[name]], dim=0)
+                if "lora_A" in name
+                else torch.cat([output_state_dict[name], -pissa_init_state_dict[name]], dim=1)
+            )
+        self.delete_adapter(adapter_name)
+        return tensors_lora
